+++ src/base/base.cmake
@@ -39,6 +39,7 @@ set(HEADERS_BASE
     src/base/net/stratum/strategies/FailoverStrategy.h
     src/base/net/stratum/strategies/SinglePoolStrategy.h
     src/base/net/stratum/SubmitResult.h
+    src/base/net/stratum/Url.h
     src/base/net/tools/RecvBuf.h
     src/base/net/tools/Storage.h
     src/base/tools/Arguments.h
@@ -75,6 +76,7 @@ set(SOURCES_BASE
     src/base/net/stratum/Pools.cpp
     src/base/net/stratum/strategies/FailoverStrategy.cpp
     src/base/net/stratum/strategies/SinglePoolStrategy.cpp
+    src/base/net/stratum/Url.cpp
     src/base/tools/Arguments.cpp
     src/base/tools/Buffer.cpp
     src/base/tools/String.cpp
@@ -95,7 +97,14 @@ elseif (APPLE)
 else()
     set(SOURCES_OS
         src/base/io/json/Json_unix.cpp
-        src/base/kernel//Platform_unix.cpp
+        src/base/kernel/Platform_unix.cpp
+        )
+endif()
+
+
+if (WITH_HWLOC)
+    list(APPEND SOURCES_OS
+        src/base/kernel/Platform_hwloc.cpp
         )
 endif()
 
@@ -127,6 +136,7 @@ if (WITH_HTTP)
         src/base/net/http/HttpResponse.h
         src/base/net/http/HttpServer.h
         src/base/net/stratum/DaemonClient.h
+        src/base/net/stratum/SelfSelectClient.h
         src/base/net/tools/TcpServer.h
         )
 
@@ -142,6 +152,7 @@ if (WITH_HTTP)
         src/base/net/http/HttpResponse.cpp
         src/base/net/http/HttpServer.cpp
         src/base/net/stratum/DaemonClient.cpp
+        src/base/net/stratum/SelfSelectClient.cpp
         src/base/net/tools/TcpServer.cpp
         )