+++ src/core/config/Config.cpp
@@ -23,9 +23,9 @@
  */
 
 #include <algorithm>
-#include <string.h>
+#include <cstring>
 #include <uv.h>
-#include <inttypes.h>
+#include <cinttypes>
 
 
 #include "backend/cpu/Cpu.h"
@@ -48,10 +48,14 @@
 #endif
 
 
+#ifdef XMRIG_FEATURE_CUDA
+#   include "backend/cuda/CudaConfig.h"
+#endif
+
+
 namespace xmrig {
 
-static const char *kCPU                  = "cpu";
-static constexpr const uint32_t kVersion = 1;
+static const char *kCPU     = "cpu";
 
 #ifdef XMRIG_ALGO_RANDOMX
 static const char *kRandomX = "randomx";
@@ -59,6 +63,15 @@ static const char *kRandomX = "randomx";
 
 #ifdef XMRIG_FEATURE_OPENCL
 static const char *kOcl     = "opencl";
+#endif
+
+#ifdef XMRIG_FEATURE_CUDA
+static const char *kCuda    = "cuda";
+#endif
+
+
+#if defined(XMRIG_FEATURE_NVML)
+static const char *kHealthPrintTime = "health-print-time";
 #endif
 
 
@@ -74,12 +87,20 @@ public:
 #   ifdef XMRIG_FEATURE_OPENCL
     OclConfig cl;
 #   endif
+
+#   ifdef XMRIG_FEATURE_CUDA
+    CudaConfig cuda;
+#   endif
+
+#   if defined(XMRIG_FEATURE_NVML)
+    uint32_t healthPrintTime = 60;
+#   endif
 };
 
 }
 
 
-xmrig::Config::Config() : BaseConfig(),
+xmrig::Config::Config() :
     d_ptr(new ConfigPrivate())
 {
 }
@@ -105,10 +126,26 @@ const xmrig::OclConfig &xmrig::Config::cl() const
 #endif
 
 
+#ifdef XMRIG_FEATURE_CUDA
+const xmrig::CudaConfig &xmrig::Config::cuda() const
+{
+    return d_ptr->cuda;
+}
+#endif
+
+
 #ifdef XMRIG_ALGO_RANDOMX
 const xmrig::RxConfig &xmrig::Config::rx() const
 {
     return d_ptr->rx;
+}
+#endif
+
+
+#if defined(XMRIG_FEATURE_NVML)
+uint32_t xmrig::Config::healthPrintTime() const
+{
+    return d_ptr->healthPrintTime;
 }
 #endif
 
@@ -119,12 +156,14 @@ bool xmrig::Config::isShouldSave() const
         return false;
     }
 
-    if (version() < kVersion) {
-        return true;
-    }
-
 #   ifdef XMRIG_FEATURE_OPENCL
     if (cl().isShouldSave()) {
+        return true;
+    }
+#   endif
+
+#   ifdef XMRIG_FEATURE_CUDA
+    if (cuda().isShouldSave()) {
         return true;
     }
 #   endif
@@ -139,7 +178,7 @@ bool xmrig::Config::read(const IJsonReader &reader, const char *fileName)
         return false;
     }
 
-    d_ptr->cpu.read(reader.getValue(kCPU), version());
+    d_ptr->cpu.read(reader.getValue(kCPU));
 
 #   ifdef XMRIG_ALGO_RANDOMX
     if (!d_ptr->rx.read(reader.getValue(kRandomX))) {
@@ -149,6 +188,14 @@ bool xmrig::Config::read(const IJsonReader &reader, const char *fileName)
 
 #   ifdef XMRIG_FEATURE_OPENCL
     d_ptr->cl.read(reader.getValue(kOcl));
+#   endif
+
+#   ifdef XMRIG_FEATURE_CUDA
+    d_ptr->cuda.read(reader.getValue(kCuda));
+#   endif
+
+#   ifdef XMRIG_FEATURE_NVML
+    d_ptr->healthPrintTime = reader.getUint(kHealthPrintTime, d_ptr->healthPrintTime);
 #   endif
 
     return true;
@@ -170,7 +217,6 @@ void xmrig::Config::getJSON(rapidjson::Document &doc) const
     doc.AddMember("api",               api, allocator);
     doc.AddMember("http",              m_http.toJSON(doc), allocator);
     doc.AddMember("autosave",          isAutoSave(), allocator);
-    doc.AddMember("version",           kVersion, allocator);
     doc.AddMember("background",        isBackground(), allocator);
     doc.AddMember("colors",            Log::colors, allocator);
 
@@ -184,14 +230,21 @@ void xmrig::Config::getJSON(rapidjson::Document &doc) const
     doc.AddMember(StringRef(kOcl),     cl().toJSON(doc), allocator);
 #   endif
 
-    doc.AddMember("donate-level",      m_pools.donateLevel(), allocator);
-    doc.AddMember("donate-over-proxy", m_pools.proxyDonate(), allocator);
-    doc.AddMember("log-file",          m_logFile.toJSON(), allocator);
-    doc.AddMember("pools",             m_pools.toJSON(doc), allocator);
-    doc.AddMember("print-time",        printTime(), allocator);
-    doc.AddMember("retries",           m_pools.retries(), allocator);
-    doc.AddMember("retry-pause",       m_pools.retryPause(), allocator);
-    doc.AddMember("syslog",            isSyslog(), allocator);
-    doc.AddMember("user-agent",        m_userAgent.toJSON(), allocator);
-    doc.AddMember("watch",             m_watch, allocator);
+#   ifdef XMRIG_FEATURE_CUDA
+    doc.AddMember(StringRef(kCuda),    cuda().toJSON(doc), allocator);
+#   endif
+
+    doc.AddMember("donate-level",               m_pools.donateLevel(), allocator);
+    doc.AddMember("donate-over-proxy",          m_pools.proxyDonate(), allocator);
+    doc.AddMember("log-file",                   m_logFile.toJSON(), allocator);
+    doc.AddMember("pools",                      m_pools.toJSON(doc), allocator);
+    doc.AddMember("print-time",                 printTime(), allocator);
+#   if defined(XMRIG_FEATURE_NVML)
+    doc.AddMember(StringRef(kHealthPrintTime),  healthPrintTime(), allocator);
+#   endif
+    doc.AddMember("retries",                    m_pools.retries(), allocator);
+    doc.AddMember("retry-pause",                m_pools.retryPause(), allocator);
+    doc.AddMember("syslog",                     isSyslog(), allocator);
+    doc.AddMember("user-agent",                 m_userAgent.toJSON(), allocator);
+    doc.AddMember("watch",                      m_watch, allocator);
 }