+++ src/core/config/Config_default.h
@@ -57,6 +57,7 @@ R"===(
         "huge-pages": true,
         "hw-aes": null,
         "priority": null,
+        "memory-pool": false,
         "max-threads-hint": 100,
         "asm": true,
         "argon2-impl": null,
@@ -68,6 +69,13 @@ R"===(
         "cache": true,
         "loader": null,
         "platform": "AMD",
+        "cn/0": false,
+        "cn-lite/0": false
+    },
+    "cuda": {
+        "enabled": false,
+        "loader": null,
+        "nvml": true,
         "cn/0": false,
         "cn-lite/0": false
     },
@@ -87,10 +95,12 @@ R"===(
             "enabled": true,
             "tls": false,
             "tls-fingerprint": null,
-            "daemon": false
+            "daemon": false,
+            "self-select": null
         }
     ],
     "print-time": 60,
+    "health-print-time": 60,
     "retries": 5,
     "retry-pause": 5,
     "syslog": false,